Fix: devide files
